Re-arranging sections in the General & Container ServiceFormSection
This commit:
- Separates the advanced section from ContainerServiceFormSection to a new
  component : ContainerServiceFormAdvancedSection
- Re-arranges advanced section fragments in the GeneralServiceFormSection
